:muscle: separate shell mod
