+++ src/generate.ts
@@ -7,7 +7,8 @@ import type {
   Shell,
   Source,
 } from "./schema.ts";
-import { Converter, getShell } from "./shell.ts";
+import { Converter } from "./shell/type.ts";
+import { getShell } from "./shell/mod.ts";
 import { err, ok, Result } from "./deps.ts";
 import { GraphNode, topologicalSort } from "./sort.ts";
