+++ src/shell/zsh.ts
@@ -1,0 +1,6 @@
+import type { Converter } from "./type.ts";
+import { bash } from "./bash.ts";
+
+export const zsh: Converter = {
+  ...bash,
+} as const;
